+++ lib/pacemaker/Makefile.am
@@ -30,6 +30,7 @@ libpacemaker_la_LIBADD	= $(top_builddir)/lib/pengine/libpe_status.la \
 libpacemaker_la_SOURCES	=
 libpacemaker_la_SOURCES += pcmk_cluster_queries.c
 libpacemaker_la_SOURCES += pcmk_fence.c
+libpacemaker_la_SOURCES += pcmk_graph_consumer.c
 libpacemaker_la_SOURCES += pcmk_graph_logging.c
 libpacemaker_la_SOURCES += pcmk_graph_producer.c
 libpacemaker_la_SOURCES += pcmk_output.c
@@ -47,6 +48,3 @@ libpacemaker_la_SOURCES += pcmk_sched_promotable.c
 libpacemaker_la_SOURCES += pcmk_sched_transition.c
 libpacemaker_la_SOURCES += pcmk_sched_utilization.c
 libpacemaker_la_SOURCES += pcmk_sched_utils.c
-libpacemaker_la_SOURCES += pcmk_trans_graph.c
-libpacemaker_la_SOURCES += pcmk_trans_unpack.c
-libpacemaker_la_SOURCES += pcmk_trans_utils.c
